refactor: rename functions subpackage into internals
