+++ quill/__init__.py
@@ -1,1 +1,1 @@
-from . import classes, functions, nn, optim, utils
+from . import classes, internals, nn, optim, utils
