+++ quill/nn/_conv2d.py
@@ -4,7 +4,7 @@ from cupy.random import rand
 from . import Module
 from .functional import addbias, conv3d, sum, stack, unstack
 from ..classes import Tensor
-from ..functions import expr_check, len_check, type_check
+from ..internals import expr_check, len_check, type_check
 
 class Conv2d(Module):
