+++ quill/nn/_linear.py
@@ -3,7 +3,7 @@ from cupy.random import rand
 from . import Module
 from .functional import addbias, expand_dims, matmul, squeeze
 from ..classes import Tensor
-from ..functions import expr_check, type_check
+from ..internals import expr_check, type_check
 
 class Linear(Module):
