+++ quill/nn/_maxpool2d.py
@@ -3,7 +3,7 @@ from collections.abc import Sequence
 from . import Module
 from .functional import maxpool3d
 from ..classes import Tensor
-from ..functions import expr_check, len_check, type_check
+from ..internals import expr_check, len_check, type_check
 
 class MaxPool2d(Module):
