+++ quill/nn/functional/_avgpool3d.py
@@ -3,7 +3,7 @@ from cupy import mean, reciprocal, zeros
 from numpy import ndarray
 
 from ...classes import Tensor
-from ...functions import expr_check, len_check, type_check
+from ...internals import expr_check, len_check, type_check
 
 def avgpool3d(input_tensor: Tensor, kernel_size: int | Sequence[int], stride: int | Sequence[int] | None = None) -> Tensor:
