+++ quill/nn/functional/_concatenate.py
@@ -3,7 +3,7 @@ from cupy import concatenate as _concatenate, split
 from numpy import ndarray
 
 from ...classes import Tensor
-from ...functions import type_check
+from ...internals import type_check
 
 def concatenate(tensors: Sequence[Tensor], axis: int = 0) -> Tensor:
