+++ quill/nn/functional/_expand_dims.py
@@ -1,7 +1,7 @@
 from cupy import expand_dims as _expand_dims, squeeze
 
 from ...classes import Tensor
-from ...functions import type_check
+from ...internals import type_check
 
 def expand_dims(x: Tensor, axis: int) -> Tensor:
