+++ quill/nn/functional/_flatten.py
@@ -1,5 +1,5 @@
 from ...classes import Tensor
-from ...functions import type_check
+from ...internals import type_check
 
 def flatten(x: Tensor) -> Tensor:
