+++ quill/nn/functional/_maxpool3d.py
@@ -3,7 +3,7 @@ from cupy import amax, argmax, unravel_index, zeros
 from numpy import ndarray
 
 from ...classes import Tensor
-from ...functions import expr_check, len_check, type_check
+from ...internals import expr_check, len_check, type_check
 
 def maxpool3d(input_tensor: Tensor, kernel_size: int | Sequence[int], stride: int | Sequence[int] | None = None) -> Tensor:
