+++ quill/nn/functional/_relu.py
@@ -1,7 +1,7 @@
 from cupy import maximum
 
 from ...classes import Tensor
-from ...functions import type_check
+from ...internals import type_check
 
 def relu(x: Tensor) -> Tensor:
