+++ quill/nn/functional/_reshape.py
@@ -1,7 +1,7 @@
 from collections.abc import Sequence
 
 from ...classes import Tensor
-from ...functions import type_check
+from ...internals import type_check
 
 def reshape(x: Tensor, newshape: Sequence[int]) -> Tensor:
