+++ quill/nn/functional/_stack.py
@@ -3,7 +3,7 @@ from cupy import split, squeeze, stack as _stack
 from numpy import ndarray
 
 from ...classes import Tensor
-from ...functions import type_check
+from ...internals import type_check
 
 def stack(tensors: Sequence[Tensor], axis: int = 0) -> Tensor:
