+++ quill/nn/functional/_sum.py
@@ -1,7 +1,7 @@
 from cupy import expand_dims, sum as _sum, repeat
 
 from ...classes import Tensor
-from ...functions import type_check
+from ...internals import type_check
 
 def sum(x: Tensor, axis: int = 0) -> Tensor:
