+++ quill/nn/functional/_tanh.py
@@ -2,7 +2,7 @@ from cupy import exp
 from numpy import ndarray
 
 from ...classes import Tensor
-from ...functions import type_check
+from ...internals import type_check
 
 def tanh(x: Tensor) -> Tensor:
