+++ quill/nn/functional/_unstack.py
@@ -1,7 +1,7 @@
 from cupy import split, squeeze
 
 from ...classes import Tensor
-from ...functions import type_check
+from ...internals import type_check
 
 def unstack(tensor: Tensor, axis: int = 0) -> list[Tensor]:
